[!] Improve AuthToken implementation
The AuthToken now provides interfaces and an abstract class.

Changed:
- New base class `AbstractAuthToken` implements `AuthTokenInterface`
- Child class `AuthToken` implements `AuthTokenCookieInterface`
- Method `AbstractAuthenticator::authenticateByToken()` to ensure token supports cookies

Deprecated:
- `AuthTokenMetadata` option "cookie_name" in favour of "token_name"
- `AuthTokenMetadata` option "cookie_duration" in favour of "token_duration"
