+++ metadata/charcoal/user/auth-token.json
@@ -1,26 +1,4 @@
 {
-    "properties": {
-        "ident": {
-            "type": "string",
-            "unique": true
-        },
-        "token": {
-            "type": "string"
-        },
-        "user_id": {
-            "type": "string"
-        },
-        "expiry": {
-            "type": "date-time"
-        },
-        "created": {
-            "type": "date-time"
-        },
-        "last_modified": {
-            "type": "date-time"
-        }
-    },
-
     "sources": {
         "default": {
             "table": "charcoal_user_auth_tokens"
